# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/usuarios/{id}": {
      "get": {
        "operationId": "UsuariosController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get user by id",
        "tags": [
          "Usuarios"
        ]
      },
      "put": {
        "operationId": "UsuariosController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update your user",
        "tags": [
          "Usuarios"
        ]
      },
      "delete": {
        "operationId": "UsuariosController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a user (admin)",
        "tags": [
          "Usuarios"
        ]
      }
    },
    "/usuarios": {
      "get": {
        "operationId": "UsuariosController_getUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicUserDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List users (admin)",
        "tags": [
          "Usuarios"
        ]
      },
      "post": {
        "operationId": "UsuariosController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserDto"
                }
              }
            }
          }
        },
        "summary": "Create a new user",
        "tags": [
          "Usuarios"
        ]
      }
    },
    "/postulaciones/analizar": {
      "post": {
        "operationId": "PostulacionesController_analyze",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzePostulacionDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Preview of the URL"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Analyze a job posting URL and return preview metadata",
        "tags": [
          "postulaciones"
        ]
      }
    },
    "/postulaciones": {
      "post": {
        "operationId": "PostulacionesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostulacionDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Postulacion created"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new postulacion for the authenticated user",
        "tags": [
          "postulaciones"
        ]
      },
      "get": {
        "operationId": "PostulacionesController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "ENVIADO",
              "type": "string"
            }
          },
          {
            "name": "isArchived",
            "required": false,
            "in": "query",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "List of postulaciones"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List postulaciones for the authenticated user",
        "tags": [
          "postulaciones"
        ]
      }
    },
    "/postulaciones/{id}": {
      "get": {
        "operationId": "PostulacionesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Postulacion returned"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get postulacion by id",
        "tags": [
          "postulaciones"
        ]
      },
      "put": {
        "operationId": "PostulacionesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostulacionDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Postulacion updated"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a postulacion",
        "tags": [
          "postulaciones"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsuarioDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered and token returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResultDto"
                }
              }
            }
          }
        },
        "summary": "Register a new user and return token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns access token and user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResultDto"
                }
              }
            }
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "auth"
        ]
      }
    }
  },
  "info": {
    "title": "Postulado API",
    "description": "API para la gestión de usuarios y postulaciones",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "9a1b2c3d-..."
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "example": "Rodrigo"
          },
          "role": {
            "type": "string",
            "example": "USER"
          },
          "birthDate": {
            "type": "string",
            "example": "1998-05-20T00:00:00.000Z"
          },
          "countryCode": {
            "type": "string",
            "example": "AR"
          },
          "createdAt": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "birthDate",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateUsuarioDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "Secret123",
            "writeOnly": true
          },
          "name": {
            "type": "string",
            "example": "Rodrigo"
          },
          "birthDate": {
            "type": "string",
            "example": "1998-05-20"
          },
          "countryCode": {
            "type": "string",
            "example": "AR"
          }
        },
        "required": [
          "email",
          "password",
          "birthDate"
        ]
      },
      "UpdateUsuarioDto": {
        "type": "object",
        "properties": {}
      },
      "AnalyzePostulacionDto": {
        "type": "object",
        "properties": {}
      },
      "CreatePostulacionDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://empresa.com/oferta/123"
          },
          "title": {
            "type": "string",
            "example": "Frontend Developer"
          },
          "description": {
            "type": "string",
            "example": "Fullstack role at ACME"
          }
        },
        "required": [
          "url"
        ]
      },
      "UpdatePostulacionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "New title"
          },
          "description": {
            "type": "string",
            "example": "Updated description"
          },
          "status": {
            "type": "string",
            "example": "ENTREVISTA"
          },
          "isArchived": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "AuthResultDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          }
        },
        "required": [
          "access_token",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "Secret123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      }
    }
  }
}